Refactor common code to common
