+++ src/CMakeLists.txt
@@ -1,2 +1,3 @@
 add_subdirectory(sequential)
 add_subdirectory(parallel)
+add_subdirectory(traces)
